+++ themes/positron/js/electro.js
@@ -9,6 +9,12 @@ var App = App || {}; // Create namespace
     App.SEARCH_RESULT_SNIPPET_MAX_LEN = 200;
 
     App.main = () => {
+        // Show version
+        console.log('Built with Electro ' + App.globalConfig.electroVersion);
+
+        const pageId = App.getUrlValue('pageId');
+        const headingId = App.getUrlValue('headingId');
+
         // -----------------------
         // menu-tree: Find all items (spans) in all menu-tree(s)
         // -----------------------
@@ -27,46 +33,22 @@ var App = App || {}; // Create namespace
         for (let span of App.state.allMenuSpans) {
             span.addEventListener("click", function () {
                 App.onClickMenuItem(this);
-                // // Unselect all items in all menu-tree(s)
-                // for (let old_span of App.state.allMenuSpans) {
-                //     old_span.classList.remove("selected");
-                //     old_span.classList.remove("navigating");
-                // }
-
-                // const pageId = this.dataset.documentName;
-                // const targetHeadingId = this.dataset.targetHeadingId;
-                
-                // // -----------------------------
-                // // Make the target page visible
-                // // -----------------------------
-                // App.showPage(pageId);
-
-                // // ----------------------------
-                // // Scroll to the target heading (or to top if no HeadingId exists)
-                // // -----------------------------
-                // App.scrollToHash(targetHeadingId);
-
-                // // -----------------------------
-                // // Select the clicked menu item
-                // // -----------------------------
-                // this.classList.add("selected");
-
-                // // If in responsive narrow-screen, re-hide the menu.
-                // document.getElementById("sidebar-container").classList.remove("force-show");
             });
         }
 
         // -----------------------
         // menu-tree: Select the current document (at page load)
         // -----------------------
-        for (let span of App.state.allMenuSpans) {
-            if (
-                (span.id == "menuitem_doc_" + App.globalConfig.currentDocumentName)
-                && (span.classList.contains("level-0"))
-            ){
-                span.classList.add("selected");
+        if(App.globalConfig.singleFile){
+            if (pageId){
+                App.selectMenuItem(pageId, headingId);
             }
         }
+        else {
+            // Static site
+            App.selectMenuItem(App.globalConfig.currentDocumentName, null);
+        }
+
 
         // -----------------------
         // menu-tree: Set caret (child folding) click handler
@@ -142,8 +124,6 @@ var App = App || {}; // Create namespace
         }
         
         // Navigate to search result location (if requested in url)
-        const pageId = App.getUrlValue('pageId');
-        const headingId = App.getUrlValue('headingId');
         if(pageId){
             console.log('Going to location: pageId:' + pageId + ' headingId:' + headingId);
             App.showPage(pageId);
@@ -164,9 +144,47 @@ var App = App || {}; // Create namespace
         const main_element = document.getElementsByClassName("main-container")[0];
         main_element.addEventListener('touchmove', App.onMainTouchMove, false);
 
-        
-        // Show version
-        console.log('Built with Electro ' + App.globalConfig.electroVersion);
+    };
+
+    // Highlight the sidebar menu item associated with documentName and (if non-null) headingName.
+    App.selectMenuItem = (documentName, headingName) => {
+        var spanDocument = null;
+        var spanHeading = null;
+        console.log('selectMenuItem(): ' + documentName + " :: " + headingName);
+        for (let span of App.state.allMenuSpans) {
+            if (
+                (span.id == "menuitem_doc_" + documentName)
+                && (span.classList.contains("level-0"))
+            ){
+                // Found document menu item
+                spanDocument = span;
+            }
+            if (
+                headingName
+                && (span.id == "menuitem_doc_" + documentName)
+                && (span.dataset.targetHeadingId == headingName)
+            ){
+                // Found heading menu item
+                spanHeading = span;
+            }
+        }
+        if (!spanDocument){
+            console.log('   document ' + documentName + ' not found.');
+            return;
+        }
+        if(!spanHeading){
+            console.log('   heading ' + headingName + ' not found. Selecting document menu item.');
+            spanDocument.classList.add("selected");
+        } else{
+            console.log('   heading ' + headingName + ' found. Selecting heading menu item.');
+            // Unfold the ul associated with the document menu item (so that the heading menu
+            // item will be visible).
+            var list = spanDocument.parentElement.getElementsByTagName('ul')[0];
+            list.classList.add("active");
+
+            // Select the heading menu item
+            spanHeading.classList.add("selected");
+        }
     };
 
     App.onClickMenuItem = (self) => {
@@ -290,6 +308,7 @@ var App = App || {}; // Create namespace
                 return decodeURI(KeyValuePair[1]);
             }
         }
+        return null;
     };
 
     App.onSearch = () => {